transport/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mysql-connector-python==8.0.31
Pillow==9.2.0
protobuf==3.20.1
psycopg2-binary==2.9.3
PyJWT==2.6.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
